Support showing local assets in Asset Browsers ("Local" repository)
Adds a new repository "Local" (will probably rename this in a bit, together
with other UI improvements).

The asset browser shows the assets in the current Blender session then. Note
that there's a difference between the data saved in the file, and the current
session's data - the file may have unsaved changes.
So this new repository acts on the current runtime "main" data, not on a file
stored on disk.

Note that the Asset Browser still has to refreshed or reopened when
creating/deleting assets, so that the changes are reflected.