Add Preferences options to set up custom asset repositories
Can be found under the File Paths category. There already is a default
repository, called "Default" and pointing to the documents or home
directory. For most Linux users it will be ~/assets.blend, for Windows
users it will be the Documents directory.

Note that this is only the Preference part of this. The Asset-Browser
doesn't use this yet.

For now, asset repositories will be single .blend files. There are some
design questions regarding how the Asset Browser would display a
directory structure as repository.