Fix T76595: Indicate the Active Keyframe in Graph Editor
Currently there is a panel that says "Active Keyframe" for numerically editing one keyframe's values, but in the code there is no concept of the "active keyframe."

This patch adds an "active keyframe index" to each FCurve, and displays it with a theme color for the active vertex (which didn't exist before) if the FCurve is active.

{F8536092 size=full}

The active keyframe is not currently set for select operations other than basic click-select, which mirrors the behavior in the 3D view.

Reviewed By: Severin, looch

Maniphest Tasks: T76595

Differential Revision: https://developer.blender.org/D7737