Cycles: add support for BVH refit in OptiX
In order to achieve that we cache the OptiX datat handles in the Geometry structure, and we do not free the memory from the original BVH build. When building the BVH for a refit, we simply tell OptiX that the build is supposed to be an update. For this, we should also tell OptiX that the initial build should allow for updates, however this is only supported in viewport renders for now.

I have not done any profile to check how fast things are here, however it seems that the bottleneck is still data transfer.

Differential Revision: https://developer.blender.org/D9353