Fix T81874: Box Select Keyframes doesnt set an active keyframe
Set the active keyframe on box/circle/lasso select in the graph editor.

For every FCurve that is influenced by the select operator that doesn't
have an active keyframe yet, the right-most keyframe is marked as the
active Keyframe.

Note that what is seen as the active keyframe for the graph editor also
depends on the active FCurve, which is not changed by box/circle/lasso
select.