Support storing asset repository data in files
Given that a repository can be built up from a single, or multiple
.blends, it's not clear where the repository data should live. It should
still be possible to easily share a .blend file to share assets,
including relevant meta-data that is not solely stored on the asset
level.
E.g. if you want to share a single pose library, that should be
doable by sharing a single .blend. So the .blend needs to have the pose
library meta-data, i.e. the name of the library and which poses belong
to it. The current design forsees that the Pose Library is implemented
as a "cataloge" (aka "smart-filter", to be determined). These should be
stored per asset category in the repository.

So this commit works towards the following design idea:
Each .blend file with assets can additionally store repository
meta-data. When reading a repository, we don't only read the asset
meta-data but also the repository meta-data in each .blend file. The
union over all the .blend files is the total repository meta-data the
asset browser can display.
That way each .blend file still is an isolated unit that doesn't
necessarily depend on the existence of other .blend files (or an index)
in the same repository.

This commit implements the storage framework of the repository meta-data.

I also added a temporary define to help spotting the related changes in
future.