USD improter:  simplify mesh reading.
The original implementation of USDMeshReader::read_mesh()
closely followed AbcMeshReader::read_mesh(), which was
designed to be invoked by a mesh modifier.  However,
since the USD importer isn't applying a mesh modifier
yet, this made the code unnecessarily complex.  For now
I've refactored this function to simply return a new
mesh, and the additional logic for modifiers can be
restored in the future when it's needed.