Revert "Codesign: Versioning code to support older branches"
This reverts commit 9d172f007eeb5ade5ddad03e1f71fa4c373855f8.

Got a second thought and remembered why it was not done in the first place.
The issue here is that the server needs to communicate codesign result back
and that must happen within the new protocol. So if the client talks old
protocol it is possible to receieve data from it, but is not possible to
communicate result back to it.