Make object preview generation thread-safe
When creating the preview, the object is now deep-copied so the preview
generation doesn't have to be done on the main thread, or block it.

There's a failing assert for objects with particle settings, I think the assert
needs to be corrected but I'll leave this aside for later. Disabled it for now.