Fix wrong assets displayed if repository is invalid
If the repository did not point to a valid .blend, all .blend files in the
parent directory were read, and all their assets shown. While this may be a
cool side effect, it's not wanted by design and would break when dragging in
assets.

The followup commit will add a hint text to the Asset Browser UI if the
repository is invalid.