Support reading all asset meta-data from external repositories
When displaying a repository that is not the "Current File" one, we can now
show all the meta-data of its assets in the UI. Editing is currently possible,
but changes will be lost on reload of the repository. Editing should simply not
be disabled for now.

Note that this also works for custom properties. So assets can store custom
properties, and the Asset Browser can show them, even if they are stored in an
external repository.
One important thing is that custom asset properties must not store pointers to
other data-blocks, but the BPY should forbid this already for assets.

Addresses T82860.