Support custom preview images for assets
Assets types that already support previews (objects, collections, images,
textures, materials, brushes, lights, worlds, scenes and screens) now support
custom previews too.
To assign a preview, the Asset Browser sidebar shows a "Load Custom Preview"
button. It opens a File Browser to select an image. Once an image is selected,
it overrides whatever preview was previously set. No reference to the image is
created, it basically copies the image to the data-block preview.

The button looks a bit out-of-place currently, that will be improved later.