Turn file previews into icons to allow access from BPY
That way we can expose an icon-id that scripts can use to reference the icon.
The Python API for file entries has the icon-id exposed.

To get this to work I had to make parts of the icon system thread safe. I added
a bunch of spin locks, but as a whole it's rather weak.

Part of T82977.