Asset Browser: Support dragging assets and appending on drop
Idea is simple: When dragging an asset, store the source file-path and
data-block name and when dropping, append the data-block. It uses existing drop
operators, but the function to get the dropped ID is replaced with one that
returns the local ID, or, in case of an external asset, appends the ID first.

Note that I expect this to be a short-lived change. A refactor like D4071 is
needed to make the drag & drop system more future proof for assets and other
use cases.

Differential Revision: https://developer.blender.org/D9721