UI: Tracking: Remove operator options exposed as props
In the movie clip editor, some options of operators are exposed as
properties and are shown in the toolbar.
Other areas of Blender do not do this making the movie clip editor break
paradigm.
This change is needed to migrate the current toolbar to a new active
tool based toolbar

Differential Revision: https://developer.blender.org/D9805