Directly select animation curves in the graph editor
Hi!

In Blender, in the graph editor, it is for now impossible to select a curve directly without selecting a keyframe of that curve.
This can sometimes be a bit frustrating, especially for people coming from softwares where this is possible. Being able to select curves directly allows to hide or lock them a bit faster, for instance.

This patch enables the user to **directly select animation curves with the usual selection operators**. When doing a selection, the usual keyframe selection is first performed, and if no keyframe could be found, then it tries selecting the curve. This is done on a per-curve basis (per-channel), which means that the user may select both a keyframe for a curve A, and a curve B with no keyframe. I found it to be more intuitive, but maybe this is something that should change.

The selection of the curves is done by sampling them in the interval defined by the xmin and the xmax of the selection area, and checking if those points are inside of this area.

Selecting curve is available in the **box selection** operator, the **lasso**, and the **circle selection** operators.

{F8892684}

Reviewed By: sybren, zeddb, looch

Differential Revision: https://developer.blender.org/D8687