NLA: Strip Evaluate Held Strips Even When Not First Strip
Changes NLA strip evaluation.

A visual example of the change (with current frame between both strips):

**Before**:
| First strip extrapolation | Next strip extrapolation | Evaluated Strip
| None | Hold | Neither
| Hold or Hold_Forward| Hold | First

**Now**:
| First strip extrapolation | Next strip extrapolation | Evaluated Strip
| None | Hold | Next
| Hold or Hold_Forward | Hold | First

There isn't a concrete use-case for this new behavior. It's more to avoid the case of [None | Hold | Neither] which appears buggy.

UI-wise, the case of `Hold`/`Hold_Forward` leading to a `Hold` will draw with a slightly darker extrapolation area overlay.  Pre-extrapolation and post-extrapolation alpha is changed to the same value now too.

Part of {T82230}
This patch functionally depends on {D9942} to sanely test this patch.

Differential Revision: https://developer.blender.org/D9943