Tracking: Implement proportional editing
Disclaimer: This is a work-in-progress, to find best and optimal
workflow, user interface, defaults, and so on.

To test the new functionality:
1. Enable proportional editing.
2. Grab marker.

The expected behavior is that in this configuration the transform
will be smoothly propagated up to keyframe or up to end of tracked
segment (whatever happens sooner).

Allows to address issue of sudden jump in track when correcting
sliding marker.

Differential Revision: https://developer.blender.org/D2771