UI: Tracking: Solve panel to Sidebar
This change moves the Solve panel to Sidebar
I was thinking it might be a good idea to split up the properties,
keyframe settings to Objects, Tripod and intrinsics to camera.
However because they are interlaced I left them together.

This change is needed to migrate the current toolbar to a new active tool based toolbar

Differential Revision: https://developer.blender.org/D9808