EEVEE: Depth of field: Improve scatter pass
This makes each sprite scatter 4 pixels instead of one.

This roughly divides the render time of the effect by two and
improve the color precision.

The single pass method was creating issues on near/far CoC
discontinuities, so we reverted to a 2 pass method even if a bit
more costly (but still is 2x faster).