Build: OpenColorIO 2.0.0 support for make deps and install_deps.sh
* Tinyxml was replaced with Expat, adding a new dependency.
* Yaml-cpp is now built as a dependency on Unix, as was already done on Windows.
* Removed currently unused LCMS code, couldn't be bothered to update and test it.
* Pystring remains built as part of OCIO itself, since it has no good build system.

We currently require OpenColorIO 2.0.0 and the upcoming code changes have no
compatibility with 1.x. Compatibility could be added, but it's rather complicated
and I only want to do the work if it's really needed.

The install_deps.sh support was implemented by Jeroen, I have no tested it myself.

The Windows code was updated, but I have not tested if it actually works, it
almost certainly will not.