OpenColorIO: update processors and transforms for version 2.0
CPU processors now need to be created to do CPU processing. These are cached
internally, but the cache lookup is not fast enough to execute per pixel or
texture sample, so for performance these are now also exposed in the C API.

The C API for transform will no longer be needed afer all changes, so remove
it to simplify the API and fallback implementation.