OpenColorIO: use aces_interchange role to detect CIE XYZ values
We need standard CIE XYZ values for rendering effects like blackbody emission.
The relation to the scene linear role is based on OpenColorIO configuration.

In OpenColorIO 2.0 configs roles can no longer have the same name as color
spaces, which means our XYZ role and colorspace in the configuration give an
error.

Instead use the new standard aces_interchange role, which relates scene linear
to a known scene referred color space. Compatibility with the old XYZ role is
preserved, if the configuration file has no conflicting names.

Also includes a non-functional change to the configuraton file to use an
XYZ-to-ACES matrix instead of REC709-to-ACES, makes debugging a little easier
since the matrix is the same one we have in the code now and that is also
found easily in the ACES specs.

There is a new cie_xyz_d65_interchange role that we should add support for,
for interchange of display referred colorspace (as opposed to scene referred).
I couldn't immediately understand how that is supposed to be defined, will
probably leave that for another patch unless someone knows how.