USD importer stage traversal options.
Added new stage traversal options to filter prim import
based on purpose (guide, proxy and render) and whether
to traverse instances with instance proxies.  Also modified
the stage traversal logic to avoid a small memory leak due to
creating a reader for the pseudo root which is never used or
deleted.