USD import: convert UsdGeomImageables only.
Currently, the importer converts every node in the USD
to a Blender Empty object by default, including nodes
that aren't of UsdGeomImageable type, such as materials
and shaders.  This can unnecessarily bloat the Blender
scene with nodes that are not immediately useful.  I've
added logic to restrict conversion to nodes that are
UsdGeomImageables, and this can have a dramatic effect
in some cases.  For exmaple, the number of nodes imported
in the Attic scene is reduced from 1,025 to 238 with
this change.