Make the different weight paint code paths exclusive to each other
Before you could have the new sculpt symmetry code and the older weight
paint symmetry code active at the same time.  This would lead to users
easily trashing their weigh paint data if they were not careful when
switching between modes.

Now the specific weight paint symmetry code is an exclusive toggle so
the user can't accidentally mirror strokes and vertex groups at the same
time.  This also paves the way of supporting Y and Z symmetry in the
future for weight groups mirroring if we decide to add it in the future.
(As the UI part for it is done)

See https://developer.blender.org/D10426