Tracking: Implement previous/next keyframe scopes
The purpose of these scopes is to give more visual reference to a VFX artist:
this way the artist can (re)position the marker more accurately when it starts
to slide. Before this change it was always more of a guess work about where
exactly the marker was at the previous keyframe.

There are some weak aspects of the patch related on some rather annoying
duplicated lines about calling an utility function for each of 3 scopes.
Ideally there needs to be some sort of foreach scope utility, but it is not as
trivial as it sounds since the editors implementation code is in ED module,
which is not available by blenloader and blenkernel.

The interface aspects still needs some work. Currently it is 3 scopes placed in
one column. Maybe it needs to be a flow, or some collapsable box?

Differential Revision: https://developer.blender.org/D10567