USD Import:  crash on reading no polys.
Fixed a crash where polys were not being read for new meshes when
the MOD_MESHSEQ_READ_POLY flag isn't set.  Crash occurs because
downstream code expects polys to exist.  Modified the logic to
always read verts and polys for new meshes, regardless of the
value of the read flag.