USD Import: fix reader refcount error.
Currently, USDPrimReader reference counts aren't correctly
maintained because the readers are stored in two separate
arrays, in the USDStageReader instance and also in the
import job data, but the reference counts for the readers
is incremented only once. I changed the logic to store
the readers only in the USDStageReader. Now also deleting
the stage reader instance, fixing a small memory leak. Also,
simplified the syntax for for looping over readers, to make
the code more readable.